Move graph operations to a submodule
