+++ kgx/graph_operations/graph_merge.py
